refactor: bundle admin package
