+++ examples/getstarted/package.json
@@ -12,7 +12,7 @@
     "strapi": "strapi"
   },
   "dependencies": {
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/plugin-color-picker": "4.13.2",
     "@strapi/plugin-documentation": "4.13.2",
     "@strapi/plugin-graphql": "4.13.2",
