--- packages/core/admin/admin/src/app.js
@@ -1,9 +1,0 @@
-const config = {
-  locales: ['fr'],
-};
-const bootstrap = () => {};
-
-export default {
-  config,
-  bootstrap,
-};
