+++ packages/core/admin/package.json
@@ -19,16 +19,24 @@
       "url": "https://strapi.io"
     }
   ],
-  "main": "index.js",
+  "exports": {
+    "./strapi-admin": {
+      "source": "./admin/src/index.js",
+      "import": "./dist/admin/index.mjs",
+      "require": "./dist/admin/index.js",
+      "default": "./dist/admin/index.js"
+    },
+    "./strapi-server": {
+      "source": "./server/index.js",
+      "import": "./dist/server/index.mjs",
+      "require": "./dist/server/index.js",
+      "default": "./dist/server/index.js"
+    },
+    "./package.json": "./package.json"
+  },
   "scripts": {
-    "analyze:bundle": "ANALYZE_BUNDLE=true webpack --config webpack.config.dev.js",
-    "analyze:deps": "ANALYZE_DEPS=true webpack serve --config webpack.config.dev.js --progress profile",
-    "create:plugin-file": "node ./scripts/create-dev-plugins-file.js",
-    "develop": "yarn create:plugin-file && yarn develop:webpack",
-    "develop:webpack": "cross-env NODE_ENV=development webpack serve --config webpack.config.dev.js --progress profile",
+    "build": "strapi plugin:build --force",
     "prepublishOnly": "yarn build",
-    "build": "rimraf build && node ./scripts/build.js",
-    "build:measure": "rimraf build && cross-env MEASURE_BUILD_SPEED=true node ./scripts/build.js",
     "test": "echo \"no tests yet\"",
     "test:unit": "run -T jest",
     "test:unit:watch": "run -T jest --watch",
@@ -43,9 +51,9 @@
     "@casl/ability": "^5.4.3",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@strapi/data-transfer": "4.13.2",
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/permissions": "4.13.2",
     "@strapi/provider-audit-logs-local": "4.13.2",
     "@strapi/typescript-utils": "4.13.2",
@@ -129,6 +137,7 @@
     "yup": "0.32.9"
   },
   "devDependencies": {
+    "@strapi/strapi": "4.13.2",
     "@testing-library/dom": "9.2.0",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
