+++ packages/core/content-type-builder/package.json
@@ -30,10 +30,10 @@
   },
   "dependencies": {
     "@sindresorhus/slugify": "1.1.0",
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/generators": "4.13.2",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/utils": "4.13.2",
     "fs-extra": "10.0.0",
     "immer": "9.0.19",
