+++ packages/core/email/package.json
@@ -27,8 +27,8 @@
     "lint": "run -T eslint ."
   },
   "dependencies": {
-    "@strapi/design-system": "1.9.0",
-    "@strapi/icons": "1.9.0",
+    "@strapi/design-system": "1.10.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/provider-email-sendmail": "4.13.2",
     "@strapi/utils": "4.13.2",
     "lodash": "4.17.21",
