+++ packages/core/helper-plugin/package.json
@@ -65,8 +65,8 @@
     "@storybook/builder-vite": "7.4.0",
     "@storybook/react-vite": "7.4.0",
     "@strapi/admin-test-utils": "4.13.2",
-    "@strapi/design-system": "1.9.0",
-    "@strapi/icons": "1.9.0",
+    "@strapi/design-system": "1.10.0",
+    "@strapi/icons": "1.10.0",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
     "@types/react": "18.2.16",
@@ -91,8 +91,8 @@
     "webpack-cli": "^5.1.0"
   },
   "peerDependencies": {
-    "@strapi/design-system": "1.9.0",
-    "@strapi/icons": "1.9.0",
+    "@strapi/design-system": "1.10.0",
+    "@strapi/icons": "1.10.0",
     "react": "^17.0.0 || ^18.0.0",
     "react-dom": "^17.0.0 || ^18.0.0",
     "react-router-dom": "^5.3.4",
