+++ packages/core/upload/package.json
@@ -25,9 +25,9 @@
     "lint": "run -T eslint ."
   },
   "dependencies": {
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/provider-upload-local": "4.13.2",
     "@strapi/utils": "4.13.2",
     "axios": "1.5.0",
