+++ packages/plugins/color-picker/package.json
@@ -63,9 +63,9 @@
     "lint:front": "run -T eslint ./admin -c ./admin/.eslintrc.js"
   },
   "dependencies": {
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "prop-types": "^15.8.1",
     "react-colorful": "5.6.1",
     "react-intl": "6.4.1"
