+++ packages/plugins/documentation/package.json
@@ -28,9 +28,9 @@
     "test:front:watch": "run -T cross-env IS_EE=true jest --config ./jest.config.front.js --watchAll"
   },
   "dependencies": {
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/utils": "4.13.2",
     "bcryptjs": "2.4.3",
     "cheerio": "^1.0.0-rc.12",
