+++ packages/plugins/graphql/package.json
@@ -26,9 +26,9 @@
   "dependencies": {
     "@graphql-tools/schema": "8.5.1",
     "@graphql-tools/utils": "^8.13.1",
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/utils": "4.13.2",
     "apollo-server-core": "3.12.1",
     "apollo-server-koa": "3.10.0",
