+++ packages/plugins/i18n/package.json
@@ -30,9 +30,9 @@
     "lint": "run -T eslint ."
   },
   "dependencies": {
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0",
+    "@strapi/icons": "1.10.0",
     "@strapi/utils": "4.13.2",
     "formik": "2.4.0",
     "immer": "9.0.19",
