+++ packages/plugins/sentry/package.json
@@ -27,9 +27,9 @@
   },
   "dependencies": {
     "@sentry/node": "6.19.7",
-    "@strapi/design-system": "1.9.0",
+    "@strapi/design-system": "1.10.0",
     "@strapi/helper-plugin": "4.13.2",
-    "@strapi/icons": "1.9.0"
+    "@strapi/icons": "1.10.0"
   },
   "devDependencies": {
     "react": "^18.2.0",
